requirements-test.txt
# Testing Dependencies for Fractal Goals

# Core Testing Framework
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-watch==4.2.0

# Test Data Generation
factory-boy==3.3.0
faker==20.1.0

# HTTP Testing
requests-mock==1.11.0

# Code Quality
pytest-flake8==1.1.1
pytest-black==0.3.12

# Async Testing (if needed)
pytest-asyncio==0.21.1

# Coverage Reporting
coverage[toml]==7.3.2
